Didn't filter lecture's quizzes
On the chance that the user filters ended up filtering out all of the lecture's quizzes, assuming that the lecture was tied to a quiz, it would previously include the lectures without include it's quiz. This commit removes the lecture from the possible classes if all of it's quizzes have been filtered. (and of course keeps the lectures that had no quizzes to begin with)